refactored deploy task
