+++ packages/core/src/actions/artifacts.ts
@@ -2,8 +2,9 @@ import path from 'path'
 import * as fs from 'fs'
 
 import * as semver from 'semver'
-import { remove0x } from '@eth-optimism/core-utils'
-import { utils } from 'ethers'
+import { getChainId, remove0x } from '@eth-optimism/core-utils'
+import { ethers, utils } from 'ethers'
+import ora from 'ora'
 
 import { ChugSplashInputs, ParsedChugSplashConfig } from '../config'
 import {
@@ -12,6 +13,11 @@ import {
   SolidityStorageObj,
 } from '../languages'
 import { Integration } from '../constants'
+import {
+  createDeploymentFolderForNetwork,
+  writeDeploymentArtifact,
+  writeSnapshotId,
+} from '../utils'
 
 // TODO
 export type BuildInfo = any
@@ -388,4 +394,105 @@ https://github.com/ethereum-optimism/smock#note-on-using-smoddit`
   }
 
   return (output as any).storageLayout
+}
+
+export const getDeployedBytecode = async (
+  provider: ethers.providers.JsonRpcProvider,
+  address: string
+): Promise<string> => {
+  const deployedBytecode = await provider.getCode(address)
+  return deployedBytecode
+}
+
+export const createDeploymentArtifacts = async (
+  provider: ethers.providers.JsonRpcProvider,
+  parsedConfig: ParsedChugSplashConfig,
+  finalDeploymentTxnHash: string,
+  artifactFolder: string,
+  buildInfoFolder: string,
+  integration: Integration,
+  spinner: ora.Ora,
+  networkName: string,
+  deploymentFolderPath: string
+) => {
+  spinner.start(`Writing deployment artifacts...`)
+
+  createDeploymentFolderForNetwork(networkName, deploymentFolderPath)
+
+  // Save the snapshot ID if we're on the hardhat network.
+  if ((await getChainId(provider)) === 31337) {
+    await writeSnapshotId(provider, networkName, deploymentFolderPath)
+  }
+
+  for (const [referenceName, contractConfig] of Object.entries(
+    parsedConfig.contracts
+  )) {
+    const artifact = getContractArtifact(
+      contractConfig.contract,
+      artifactFolder,
+      integration
+    )
+    const { sourceName, contractName, bytecode, abi } = artifact
+
+    const buildInfo = await getBuildInfo(buildInfoFolder, sourceName)
+
+    const { constructorArgValues } = getConstructorArgs(
+      parsedConfig,
+      referenceName,
+      abi,
+      buildInfo.output,
+      sourceName,
+      contractName
+    )
+
+    const receipt = await provider.getTransactionReceipt(finalDeploymentTxnHash)
+
+    const metadata =
+      buildInfo.output.contracts[sourceName][contractName].metadata
+
+    const { devdoc, userdoc } =
+      typeof metadata === 'string'
+        ? JSON.parse(metadata).output
+        : metadata.output
+
+    const deploymentArtifact = {
+      contractName,
+      address: contractConfig.proxy,
+      abi,
+      transactionHash: finalDeploymentTxnHash,
+      solcInputHash: buildInfo.id,
+      receipt: {
+        ...receipt,
+        gasUsed: receipt.gasUsed.toString(),
+        cumulativeGasUsed: receipt.cumulativeGasUsed.toString(),
+        // Exclude the `effectiveGasPrice` if it's undefined, which is the case on Optimism.
+        ...(receipt.effectiveGasPrice && {
+          effectiveGasPrice: receipt.effectiveGasPrice.toString(),
+        }),
+      },
+      numDeployments: 1,
+      metadata:
+        typeof metadata === 'string' ? metadata : JSON.stringify(metadata),
+      args: constructorArgValues,
+      bytecode,
+      deployedBytecode: await provider.getCode(contractConfig.proxy),
+      devdoc,
+      userdoc,
+      storageLayout: await getStorageLayout(
+        contractConfig.contract,
+        artifactFolder,
+        buildInfoFolder,
+        integration
+      ),
+    }
+
+    writeDeploymentArtifact(
+      networkName,
+      deploymentFolderPath,
+      deploymentArtifact,
+      referenceName
+    )
+  }
+
+  spinner.succeed(`Wrote deployment artifacts.`)
 }
