+++ packages/core/src/index.ts
@@ -8,3 +8,4 @@ export * from './tasks'
 export * from './messages'
 export * from './deployments'
 export * from './execution'
+export * from './types'
