+++ packages/plugins/src/hardhat/index.ts
@@ -1,5 +1,3 @@
-export * from './artifacts'
 export * from './tasks'
 export * from './deployments'
 export * from './type-extensions'
-export * from './utils'
